# Improve environment initialization performance