Instead of just caching the declarations, cache the actual bound name determination. This helps especially with modern syntax utilizing let and const which have specific lexical scope.