@lahma 2025-01-27
## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|Modern|Prepared|Mean|Error|Allocated|
|------- |-------|-------|-------|-------:|-------|-------:|
| Old |CoreEval|False|False|4.357 ms|0.1745 ms|332.12 KB|
| **New** |	| **False** | **False** | **2.704 ms (-38%)** | **0.0972 ms** | **334.13 KB (+1%)** |
| Old |CoreEval|False|True|3.199 ms|0.1329 ms|317.23 KB|
| **New** |	| **False** | **True** | **2.692 ms (-16%)** | **0.0744 ms** | **319.12 KB (+1%)** |
| Old |CoreEval|True|False|2.958 ms|0.1400 ms|331.7 KB|
| **New** |	| **True** | **False** | **3.017 ms (+2%)** | **0.1400 ms** | **333.52 KB (+1%)** |
| Old |CoreEval|True|True|3.523 ms|0.0827 ms|316.74 KB|
| **New** |	| **True** | **True** | **2.856 ms (-19%)** | **0.1464 ms** | **319.24 KB (+1%)** |
| Old |Cube|False|False|16.167 ms|0.3189 ms|6248.8 KB|
| **New** |	| **False** | **False** | **14.641 ms (-9%)** | **0.2896 ms** | **6250.94 KB (0%)** |
| Old |Cube|False|True|15.281 ms|0.3043 ms|5910.32 KB|
| **New** |	| **False** | **True** | **14.023 ms (-8%)** | **0.2803 ms** | **5912.09 KB (0%)** |
| Old |Cube|True|False|14.407 ms|0.2856 ms|5943.43 KB|
| **New** |	| **True** | **False** | **13.795 ms (-4%)** | **0.2755 ms** | **6287.93 KB (+6%)** |
| Old |Cube|True|True|14.805 ms|0.2935 ms|5606.58 KB|
| **New** |	| **True** | **True** | **14.233 ms (-4%)** | **0.2833 ms** | **5948.69 KB (+6%)** |
| Old |ObjectArray|False|False|31.745 ms|0.6204 ms|96269.91 KB|
| **New** |	| **False** | **False** | **29.290 ms (-8%)** | **0.5845 ms** | **96269.75 KB (0%)** |
| Old |ObjectArray|False|True|31.097 ms|0.6105 ms|96222.63 KB|
| **New** |	| **False** | **True** | **29.366 ms (-6%)** | **0.5778 ms** | **96222.66 KB (0%)** |
| Old |ObjectArray|True|False|34.542 ms|0.5652 ms|111436.61 KB|
| **New** |	| **True** | **False** | **34.076 ms (-1%)** | **0.6733 ms** | **111436.97 KB (0%)** |
| Old |ObjectArray|True|True|35.085 ms|0.6905 ms|111391.07 KB|
| **New** |	| **True** | **True** | **34.187 ms (-3%)** | **0.6608 ms** | **111391.09 KB (0%)** |
| Old |ObjectRegExp|False|False|134.142 ms|2.6594 ms|155214.48 KB|
| **New** |	| **False** | **False** | **123.978 ms (-8%)** | **2.4771 ms** | **151956.14 KB (-2%)** |
| Old |ObjectRegExp|False|True|100.370 ms|2.3303 ms|149767.58 KB|
| **New** |	| **False** | **True** | **97.948 ms (-2%)** | **3.0582 ms** | **146107.54 KB (-2%)** |
| Old |ObjectRegExp|True|False|153.451 ms|4.4005 ms|169792.42 KB|
| **New** |	| **True** | **False** | **154.527 ms (+1%)** | **4.7308 ms** | **166215.32 KB (-2%)** |
| Old |ObjectRegExp|True|True|124.220 ms|5.0376 ms|163823.71 KB|
| **New** |	| **True** | **True** | **124.381 ms (0%)** | **4.7241 ms** | **167854.47 KB (+2%)** |
| Old |ObjectString|False|False|243.169 ms|5.1335 ms|1315732.68 KB|
| **New** |	| **False** | **False** | **233.837 ms (-4%)** | **4.6628 ms** | **1316088.25 KB (0%)** |
| Old |ObjectString|False|True|236.541 ms|4.9431 ms|1315773.91 KB|
| **New** |	| **False** | **True** | **231.932 ms (-2%)** | **4.6290 ms** | **1315872.11 KB (0%)** |
| Old |ObjectString|True|False|2,378.698 ms|41.9221 ms|20822066.76 KB|
| **New** |	| **True** | **False** | **2,421.905 ms (+2%)** | **25.9332 ms** | **20821556.66 KB (0%)** |
| Old |ObjectString|True|True|2,482.403 ms|28.9155 ms|20822515.88 KB|
| **New** |	| **True** | **True** | **2,441.245 ms (-2%)** | **35.2152 ms** | **20821982.17 KB (0%)** |
| Old |StringBase64|False|False|30.616 ms|0.5933 ms|2397.48 KB|
| **New** |	| **False** | **False** | **30.447 ms (-1%)** | **0.6028 ms** | **2397.65 KB (0%)** |
| Old |StringBase64|False|True|30.621 ms|0.6001 ms|2297.67 KB|
| **New** |	| **False** | **True** | **30.195 ms (-1%)** | **0.5936 ms** | **2297.98 KB (0%)** |
| Old |StringBase64|True|False|37.054 ms|0.7168 ms|15113.04 KB|
| **New** |	| **True** | **False** | **36.041 ms (-3%)** | **0.7203 ms** | **12553.55 KB (-17%)** |
| Old |StringBase64|True|True|37.573 ms|0.7400 ms|15013.4 KB|
| **New** |	| **True** | **True** | **35.502 ms (-6%)** | **0.6975 ms** | **12453.59 KB (-17%)** |


## Jint.Benchmark.StopwatchBenchmark

| **Diff**|Method|Modern|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Execute|False|191.1 ms|3.32 ms|39.72 MB|
| **New** |	| **False** | **190.9 ms (0%)** | **2.06 ms** | **39.72 MB (0%)** |
| Old |Execute_ParsedScript|False|200.0 ms|2.59 ms|39.68 MB|
| **New** |	| **False** | **195.9 ms (-2%)** | **2.96 ms** | **39.68 MB (0%)** |
| Old |Execute|True|295.7 ms|3.16 ms|270 MB|
| **New** |	| **True** | **274.7 ms (-7%)** | **2.56 ms** | **225.25 MB (-17%)** |
| Old |Execute_ParsedScript|True|302.6 ms|1.71 ms|269.96 MB|
| **New** |	| **True** | **279.9 ms (-8%)** | **3.08 ms** | **225.21 MB (-17%)** |


## Jint.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|3d-cube|109.07 ms|0.932 ms|44653.67 KB|
| **New** |	|	| **110.09 ms (+1%)** | **1.241 ms** | **44655.68 KB (0%)** |
| Old |Run|3d-morph|101.74 ms|1.290 ms|46257.51 KB|
| **New** |	|	| **97.14 ms (-5%)** | **0.746 ms** | **46257.64 KB (0%)** |
| Old |Run|3d-raytrace|107.23 ms|1.230 ms|84207.03 KB|
| **New** |	|	| **106.46 ms (-1%)** | **1.359 ms** | **84207.4 KB (0%)** |
| Old |Run|access-binary-trees|59.61 ms|1.153 ms|57318.73 KB|
| **New** |	|	| **59.66 ms (0%)** | **0.781 ms** | **57318.98 KB (0%)** |
| Old |Run|access-fannkuch|272.53 ms|2.709 ms|103.81 KB|
| **New** |	|	| **278.62 ms (+2%)** | **2.202 ms** | **104.43 KB (+1%)** |
| Old |Run|access-nbody|117.46 ms|1.749 ms|53116.88 KB|
| **New** |	|	| **118.29 ms (+1%)** | **1.371 ms** | **53114.44 KB (0%)** |
| Old |Run|access-nsieve|86.35 ms|0.935 ms|17142.44 KB|
| **New** |	|	| **82.47 ms (-4%)** | **1.356 ms** | **17142.64 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|90.32 ms|0.968 ms|56948.35 KB|
| **New** |	|	| **87.50 ms (-3%)** | **1.380 ms** | **56948.27 KB (0%)** |
| Old |Run|bitops-bits-in-byte|137.30 ms|1.406 ms|37044.51 KB|
| **New** |	|	| **134.47 ms (-2%)** | **1.534 ms** | **37044.57 KB (0%)** |
| Old |Run|bitops-bitwise-and|71.28 ms|0.580 ms|55938.96 KB|
| **New** |	|	| **69.85 ms (-2%)** | **0.844 ms** | **55938.94 KB (0%)** |
| Old |Run|bitops-nsieve-bits|140.03 ms|1.338 ms|53916.25 KB|
| **New** |	|	| **132.47 ms (-5%)** | **1.482 ms** | **53916.22 KB (0%)** |
| Old |Run|contr(...)rsive [21]|69.83 ms|1.362 ms|83183.47 KB|
| **New** |	|	| **71.09 ms (+2%)** | **1.111 ms** | **83183.65 KB (0%)** |
| Old |Run|crypto-aes|78.36 ms|0.844 ms|10332.73 KB|
| **New** |	|	| **77.05 ms (-2%)** | **0.755 ms** | **10199.86 KB (-1%)** |
| Old |Run|crypto-md5|68.33 ms|0.863 ms|77797.24 KB|
| **New** |	|	| **66.17 ms (-3%)** | **1.277 ms** | **79161.23 KB (+2%)** |
| Old |Run|crypto-sha1|64.24 ms|0.553 ms|64484.96 KB|
| **New** |	|	| **63.60 ms (-1%)** | **0.616 ms** | **64485.02 KB (0%)** |
| Old |Run|date-format-tofte|60.01 ms|0.635 ms|53047.78 KB|
| **New** |	|	| **60.03 ms (0%)** | **0.747 ms** | **53188.52 KB (0%)** |
| Old |Run|date-format-xparb|35.02 ms|0.674 ms|24992.32 KB|
| **New** |	|	| **35.36 ms (+1%)** | **0.682 ms** | **24994.5 KB (0%)** |
| Old |Run|math-cordic|192.65 ms|2.802 ms|81976 KB|
| **New** |	|	| **193.48 ms (0%)** | **1.726 ms** | **81976.16 KB (0%)** |
| Old |Run|math-partial-sums|66.79 ms|1.127 ms|49367.67 KB|
| **New** |	|	| **69.00 ms (+3%)** | **0.597 ms** | **49367.75 KB (0%)** |
| Old |Run|math-spectral-norm|76.31 ms|1.523 ms|51826.62 KB|
| **New** |	|	| **72.03 ms (-6%)** | **1.412 ms** | **51826.91 KB (0%)** |
| Old |Run|regexp-dna|88.99 ms|1.218 ms|16827.88 KB|
| **New** |	|	| **88.83 ms (0%)** | **0.863 ms** | **16828.16 KB (0%)** |
| Old |Run|string-base64|47.05 ms|0.681 ms|3156.01 KB|
| **New** |	|	| **47.24 ms (0%)** | **0.518 ms** | **3156.42 KB (0%)** |
| Old |Run|string-fasta|101.66 ms|1.939 ms|102598.56 KB|
| **New** |	|	| **103.28 ms (+2%)** | **0.955 ms** | **102598.7 KB (0%)** |
| Old |Run|string-tagcloud|47.40 ms|0.657 ms|40456.08 KB|
| **New** |	|	| **46.08 ms (-3%)** | **0.692 ms** | **40456.74 KB (0%)** |
| Old |Run|string-unpack-code|46.93 ms|0.655 ms|69834.29 KB|
| **New** |	|	| **45.30 ms (-3%)** | **0.769 ms** | **69835.09 KB (0%)** |
| Old |Run|strin(...)input [21]|40.11 ms|0.717 ms|19905.74 KB|
| **New** |	|	| **39.57 ms (-1%)** | **0.608 ms** | **19905.51 KB (0%)** |